Introduced ADC/SBC from B-Em to fix overflow bugs.
The ISO Pascal compiler fails to parse numeric literals, raising an
"Integer overflow" error by executing a BRK instruction. This error
appears to occur due to an incorrect overflow condition that was
fixed in the somewhat related B-Em source code. Thus, it makes
sense to introduce B-Em's NMOS implementation of ADC/SBC into
Elkulator to keep the two programs synchronised and to benefit from
the fixes made to B-Em.